Move input parsing stuff into a directory
