+++ src/input_parsing/base.rs
@@ -1,0 +1,11 @@
+use crossterm::event::Event;
+
+use crate::{enums::ReedlineEvent, EditMode};
+
+use super::keybindings::Keybindings;
+
+pub trait InputParser {
+    fn parse_event(&self, event: Event) -> ReedlineEvent;
+    fn update_keybindings(&mut self, keybindings: Keybindings);
+    fn edit_mode(&self) -> EditMode;
+}
