+++ src/input_parsing/mod.rs
@@ -1,0 +1,7 @@
+mod base;
+mod emacs;
+mod keybindings;
+
+pub use base::InputParser;
+pub use emacs::EmacsInputParser;
+pub use keybindings::{default_emacs_keybindings, Keybindings};
